backend/requirements.txt
fastapi
uvicorn[standard]
ccxt
httpx
python-dotenv
typer
rich
